disamiguate create args (#345)
git-svn-id: http://svn.osgeo.org/geos/trunk@2991 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ tests/unit/noding/SegmentNodeTest.cpp
@@ -55,7 +55,7 @@ namespace tut
 
         // Create coordinates sequence
         const size_t coords_size = 2;
-        CoordSeqPtr cs( factory_->create(0, coords_size) );
+        CoordSeqPtr cs( factory_->create((size_t)0, coords_size) );
 
         ensure( 0 != cs.get() );
 
@@ -103,7 +103,7 @@ namespace tut
 
         // Create coordinates sequence
         const size_t coords_size = 2;
-        CoordSeqPtr cs( factory_->create(0, coords_size) );
+        CoordSeqPtr cs( factory_->create((size_t)0, coords_size) );
 
         ensure( 0 != cs.get() );
 
@@ -145,7 +145,7 @@ namespace tut
 
         // Create coordinates sequence
         const size_t coords_size = 2;
-        CoordSeqPtr cs( factory_->create(0, coords_size) );
+        CoordSeqPtr cs( factory_->create((size_t)0, coords_size) );
 
         ensure( 0 != cs.get() );
 
@@ -187,7 +187,7 @@ namespace tut
 
         // Create coordinates sequence
         const size_t coords_size = 2;
-        CoordSeqPtr cs( factory_->create(0, coords_size) );
+        CoordSeqPtr cs( factory_->create((size_t)0, coords_size) );
 
         ensure( 0 != cs.get() );
